app/app/echo.bak/requirements.txt
aioredis==1.3.1
asgiref==3.5.2
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.2.1
Automat==20.2.0
backports.zoneinfo==0.2.1
certifi==2020.12.5
cffi==1.14.5
channels==3.0.5
channels-redis==3.4.0
chardet==3.0.4
constantly==15.1.0
cryptography==38.0.4
daphne==3.0.2
DateTimeRange==1.0.0
Django==4.0.6
django-appconf==1.0.4
django-bootstrap-modal-forms==1.5.0
django-clearcache==1.2.1
django-debug-toolbar==3.2.1
django-model-utils==4.2.0
django-redis==4.12.1
django-select2==7.6.1
django-tinymce==3.3.0
django-widget-tweaks==1.4.8
djangorestframework==3.11.0
gunicorn==20.1.0
hiredis==2.0.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
mbstrdecoder==1.0.1
msgpack==1.0.2
packaging==20.9
Pillow==9.2.0
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydicom==2.1.2
pynetdicom
pymemcache==3.5.2
pyOpenSSL==22.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
redis==3.5.3
requests==2.24.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
Twisted==22.4.0
txaio==22.2.1
typepy==1.3.0
typing_extensions==4.3.0
urllib3==1.25.11
zope.interface==5.4.0
opencv-python==4.5.5.64
pypng==0.20220715.0
numpy==1.21.6
